+++ src/models/solana-rpc-result.ts
@@ -1,11 +1,13 @@
 // This file is a workaround to generate the desired output.
 // Workaround for https://github.com/OpenAPITools/openapi-generator/issues/13417
 
+// May contain unused imports in some cases
+// @ts-expect-error openapi generator may skip this model depending on spec variant
 import type { FeeEstimateResult } from './fee-estimate-result';
 import type { GetFeaturesEnabledResult } from './get-features-enabled-result';
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import type { GetSupportedTokensItem } from './get-supported-tokens-item';
 import type { GetSupportedTokensResult } from './get-supported-tokens-result';
+// May contain unused imports in some cases
+// @ts-expect-error openapi generator may skip this model depending on spec variant
 import type { PrepareTransactionResult } from './prepare-transaction-result';
 import type { SignAndSendTransactionResult } from './sign-and-send-transaction-result';
 import type { SignTransactionResult } from './sign-transaction-result';